feat(client): Start iframed sync flow.
Add a new Channel type - the duplex channel that can send and receive messages. To decouple the send and receive mechanisms, it accepts a sender and receiver which provide concrete strategies.

The new channel abstraction can accept arbitrary messages in the receiver, the cannel will broadcast a `message` event which can be listened for. This allows reliers to initiate communication with FxA, if this becomes necessary.

I would like to move all Channels to use this abstraction and ditch the old style.

Use this to communicate in a new auth broker - the SyncWebChannelAuthenticationBroker.

Add temporary glue code to start the FxA Sync flow from within an iframe. This is going to need to be updated as we flesh the API out.

fixes #1444
fixes #2231
issue #2235